Move slow compaction test to integration
